Tami Tab One - Chicken
Tami's chicken tab is complete to a design standpoint. Needs code intervention to make the buttons work as per the document, as well as optional particle systems for the chicken, though what we have is sufficient considering this will only be on screen for five seconds. There is opportunity for additional artwork if necessary.

la la la lava
ch ch ch chicken