flashlight and websites done
text on website with a tag "FlashlightChangeable" will enable or disable text if it should or not, means that once flashlight is open it'll change out text on a website if set.
"downloading" a file opens the downloads folder after setting a bool in the scriptable object to true - Important Note: in doing this the changes made to the filedata file stay even after play mode is exited, make sure bool at the very bottom "isAvailable" is set back to false.
working on the corkboard now.